\name{schema-class}
\docType{class}
\alias{ExternalReference-class}
\alias{SchemaAttributeGroupTable-class}
\alias{SchemaAttributeTable-class}
\alias{SchemaElementTable-class}
\alias{SchemaNotationTable-class}
\alias{SchemaTypeTable-class}
\alias{libxmlTypeTable-class}
\alias{xmlSchemaAttributeGroupRef-class}
\alias{xmlSchemaAttributeRef-class}
\alias{xmlSchemaElementRef-class}
\alias{xmlSchemaNotationRef-class}
\alias{xmlSchemaRef-class}
\alias{xmlSchemaTypeRef-class}

\title{Classes for working with XML Schema}

\description{
 These are classes used when working with XML schema
and using them to validate a document or querying the
schema for its elements.
}
\section{Slots}{
 \describe{}
}
\section{Methods}{
}

\seealso{
\code{\link{xmlSchemaValidate}}
}
\keyword{classes}
