\name{schema-class}
\docType{class}
\alias{ExternalReference-class}
\alias{SchemaAttributeGroupTable-class}
\alias{SchemaAttributeTable-class}
\alias{SchemaElementTable-class}
\alias{SchemaNotationTable-class}
\alias{SchemaTypeTable-class}
\alias{libxmlTypeTable-class}
\alias{xmlSchemaAttributeGroupRef-class}
\alias{xmlSchemaAttributeRef-class}
\alias{xmlSchemaElementRef-class}
\alias{xmlSchemaNotationRef-class}
\alias{xmlSchemaRef-class}
\alias{xmlSchemaTypeRef-class}
\alias{names,xmlSchemaRef-method}
\alias{$,xmlSchemaRef-method}

\alias{names,libxmlTypeTable-method}
\alias{$,libxmlTypeTable-method}
\alias{$<-,libxmlTypeTable-method}
\alias{coerce,libxmlTypeTable,list-method}

\alias{show,XMLSchemaValidationResults-method}


\title{Classes for working with XML Schema}

\description{
 These are classes used when working with XML schema
and using them to validate a document or querying the
schema for its elements.
The basic representation is an external/native object stored in the
\code{ref} slot.
}
%\section{Slots}{ \describe{}}
%\section{Methods}{}

\seealso{
\code{\link{xmlSchemaValidate}}
}
\keyword{classes}
